refactor: move test-only helpers into Render.Internal
The five pure helpers added for the Pandoc Table arm — alignmentStyle,
colSpecsToColgroup, cellSpanAttrs, cellColumnIndices, mergeStyleKVs —
were exported from the public Render module under an "exported for
tests" comment, polluting the API surface for consumers who only want
rpBlock / renderPandocWith.

Move them to Heist.Extra.Splices.Pandoc.Render.Internal. Render
imports them privately; the test suite imports them directly. The
public Render module's exports now stop at rpBlock / renderPandocWith
/ rawNode (rawNode stays for now — pre-existing test-only export,
out of scope for this PR).
