+++ heist-extra.cabal
@@ -100,6 +100,7 @@ library
     Heist.Extra.Splices.Pandoc.Ctx
     Heist.Extra.Splices.Pandoc.Footnotes
     Heist.Extra.Splices.Pandoc.Render
+    Heist.Extra.Splices.Pandoc.Render.Internal
     Heist.Extra.Splices.Pandoc.Skylighting
     Heist.Extra.Splices.Pandoc.TaskList
     Heist.Extra.Splices.Pandoc.Texmath
